trading_system/requirements.txt
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
requests==2.31.0
schedule==1.2.1
pytz==2023.3.post1
streamlit==1.31.0
pyzmq==26.0.3
openai==1.40.3
